+++ src/library_html5.js
@@ -184,7 +184,7 @@ var LibraryJSEvents = {
       {{{ makeSetValue('varargs', 0, 'eventTypeId', 'i32') }}};
       {{{ makeSetValue('varargs', 4, 'eventData', 'i32') }}};
       {{{ makeSetValue('varargs', 8, 'userData', 'i32') }}};
-      _emscripten_async_queue_on_thread_(targetThread, {{{ cDefine('EM_FUNC_SIG_IIII') }}}, eventHandlerFunc, eventData, varargs);
+      __emscripten_call_on_thread(0, targetThread, {{{ cDefine('EM_FUNC_SIG_IIII') }}}, eventHandlerFunc, eventData, varargs);
       stackRestore(stackTop);
     },
 #endif
@@ -2859,7 +2859,7 @@ var LibraryJSEvents = {
     // these two threads will deadlock. At the moment, we'd like to consider that this kind of deadlock would be an Emscripten runtime bug, although if
     // emscripten_set_canvas_element_size() was documented to require running an event in the queue of thread that owns the OffscreenCanvas, then that might be ok.
     // (safer this way however)
-    _emscripten_async_queue_on_thread_(targetThread, {{{ cDefine('EM_PROXIED_RESIZE_OFFSCREENCANVAS') }}}, 0, targetCanvasPtr /* satellite data */, varargs);
+    __emscripten_call_on_thread(0, targetThread, {{{ cDefine('EM_PROXIED_RESIZE_OFFSCREENCANVAS') }}}, 0, targetCanvasPtr /* satellite data */, varargs);
     stackRestore(stackTop);
   },